Se añadieron DTOs y se cambio logica repos, services y entidades

+++ src/main/java/com/example/facturacion_segunda_entrega/repositories/SaleRepository.java
@@ -3,5 +3,5 @@ package com.example.facturacion_segunda_entrega.repositories;
 import com.example.facturacion_segunda_entrega.entities.Sale;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SaleRepository extends JpaRepository<Sale, Long> {
+public interface SaleRepository extends JpaRepository<Sale, Integer> {
 }
